+++ CHANGES.rst
@@ -6,7 +6,8 @@ Changelog
 - [MPP-476] Added: child's price category can depend of his school implantation.
 - [AES-1559] Fixed: remove unused field.
 - [MPPNIVA-9][MPPNIVA-10] Added: update person endpoint.
-- [MPP-478] Added: disabled key on meal registrations and meal menu endpoints
+- [MPP-478] Added: is_disabled key on meal registrations and meal menu endpoints.
+- [MPP-478] Added: filter by parent on meal registrations sent to AES/
 - Fix endpoint description.
 
 1.9.2