+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 - [MPPNIVA-9][MPPNIVA-10] Added: update person endpoint.
 - [MPP-478] Added: is_disabled key on meal registrations and meal menu endpoints.
 - [MPP-478] Added: filter by parent on meal registrations sent to AES/
+- [MPP-466] Updated: check delay and insert it in disabled in inscriptions of a child
 - Fix endpoint description.
 
 1.9.2